+++ crates/pop-cli/Cargo.toml
@@ -32,6 +32,8 @@ sp-core = { version = "30.0.0", optional = true }
 sp-weights = { version = "29.0.0", optional = true }
 
 # parachains
+pop-parachains = { version = "0.1.0", path = "../pop-parachains"}
+
 dirs = { version = "5.0", optional = true }
 indexmap = { version = "2.2", optional = true }
 reqwest = { version = "0.11", optional = true }
