+++ crates/pop-cli/src/engines/mod.rs
@@ -1,5 +1,2 @@
-pub mod generator;
-#[cfg(feature = "parachain")]
-pub mod pallet_engine;
 #[cfg(feature = "parachain")]
 pub mod parachain_engine;
