+++ crates/pop-parachains/src/generator/mod.rs
@@ -1,0 +1,2 @@
+pub mod pallet;
+pub mod parachain;
