+++ crates/pop-parachains/src/generator/pallet.rs
@@ -1,16 +1,7 @@
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
+use crate::utils::helpers::write_to_file;
 use askama::Template;
-
-use crate::helpers::write_to_file;
-
-#[derive(Template)]
-#[template(path = "base/chain_spec.templ", escape = "none")]
-pub(crate) struct ChainSpec {
-	pub(crate) token_symbol: String,
-	pub(crate) decimals: u8,
-	pub(crate) initial_endowment: String,
-}
 
 #[derive(Template)]
 #[template(path = "pallet/Cargo.templ", escape = "none")]
@@ -36,25 +27,6 @@ pub(crate) struct PalletTests {
 	pub(crate) module: String,
 }
 
-#[derive(Template)]
-#[template(path = "base/network.templ", escape = "none")]
-pub(crate) struct Network {
-	pub(crate) node: String,
-}
-
-// todo : generate directory structure
-// todo : This is only for development
-#[allow(unused)]
-pub fn generate() {
-	let cs = ChainSpec {
-		token_symbol: "DOT".to_owned(),
-		decimals: 10,
-		initial_endowment: "1u64 << 15".to_owned(),
-	};
-	let rendered = cs.render().unwrap();
-	write_to_file(Path::new("src/x.rs"), &rendered);
-}
-
 pub trait PalletItem {
 	/// Render and Write to file, root is the path to the pallet
 	fn execute(&self, root: &PathBuf) -> anyhow::Result<()>;
@@ -65,7 +37,7 @@ macro_rules! generate_pallet_item {
 		impl PalletItem for $item {
 			fn execute(&self, root: &PathBuf) -> anyhow::Result<()> {
 				let rendered = self.render()?;
-				write_to_file(&root.join($filename), &rendered);
+				let _ = write_to_file(&root.join($filename), &rendered);
 				Ok(())
 			}
 		}
