+++ crates/pop-parachains/src/utils/helpers.rs
@@ -1,0 +1,39 @@
+use anyhow::Result;
+use std::{
+	fs::{self, OpenOptions},
+	path::Path,
+};
+pub(crate) fn sanitize(target: &Path) -> Result<()> {
+	use std::io::{stdin, stdout, Write};
+	if target.exists() {
+		print!("\"{}\" folder exists. Do you want to clean it? [y/n]: ", target.display());
+		stdout().flush()?;
+
+		let mut input = String::new();
+		stdin().read_line(&mut input)?;
+
+		if input.trim().to_lowercase() == "y" {
+			fs::remove_dir_all(target)?;
+		} else {
+			return Err(anyhow::anyhow!("User aborted due to existing target folder."));
+		}
+	}
+	Ok(())
+}
+
+pub(crate) fn write_to_file<'a>(path: &Path, contents: &'a str) -> Result<()> {
+	use std::io::Write;
+	let mut file = OpenOptions::new().write(true).truncate(true).create(true).open(path).unwrap();
+	file.write_all(contents.as_bytes()).unwrap();
+	if path.extension().map_or(false, |ext| ext == "rs") {
+		let output = std::process::Command::new("rustfmt")
+			.arg(path.to_str().unwrap())
+			.output()
+			.expect("failed to execute rustfmt");
+
+		if !output.status.success() {
+			return Err(anyhow::anyhow!("rustfmt exited with non-zero status code."));
+		}
+	}
+	Ok(())
+}
