+++ crates/pop-parachains/src/utils/pallet_helpers.rs
@@ -1,0 +1,36 @@
+use std::{
+	env::current_dir,
+	fs::{self},
+	path::{Path, PathBuf},
+};
+
+/// Resolve pallet path
+/// For a template it should be `<template>/pallets/`
+/// For no path, it should just place it in the current working directory
+pub fn resolve_pallet_path(path: Option<String>) -> PathBuf {
+	use std::process;
+
+	if let Some(path) = path {
+		return Path::new(&path).to_path_buf();
+	}
+	// Check if inside a template
+	let cwd = current_dir().expect("current dir is inaccessible");
+
+	let output = process::Command::new(env!("CARGO"))
+		.arg("locate-project")
+		.arg("--workspace")
+		.arg("--message-format=plain")
+		.output()
+		.unwrap()
+		.stdout;
+	let workspace_path = Path::new(std::str::from_utf8(&output).unwrap().trim());
+	if workspace_path == Path::new("") {
+		cwd
+	} else {
+		let pallet_path = workspace_path.parent().unwrap().to_path_buf().join("pallets");
+		match fs::create_dir_all(pallet_path.clone()) {
+			Ok(_) => pallet_path,
+			Err(_) => cwd,
+		}
+	}
+}
