+++ crates/pop-parachains/templates/base/chain_spec.templ
