+++ crates/pop-parachains/templates/base/network.templ
