+++ crates/pop-parachains/templates/pallet/Cargo.templ
